refactor(core): use a structural test for Suites
Use a structural test rather than instanceof to check if an object is a
Suite. It's possible for multiple copies of the Suite class to be loaded
at once, such as when directly importing an interface or the Suite class
in tests. That will cause an instanceof check to fail, even when a suite
object is effectively a Suite.